big fix 4

+++ main.py
@@ -131,7 +131,7 @@ def main():
 
         #input_tensor = interpreter.convert_to_tensor(np.expand_dims(image_np, 0), dtype=tf.float32)
         #img = cv2.resize(cv2.cvtColor(frame, cv2.COLOR_BGR2RGB), (320, 320))
-        detections = detect_objects(interpreter, ret, 0.8)
+        detections = detect_objects(interpreter, ret.astype(np.int64), 0.8)
 
         detections['detection_classes'] = detections['detection_classes'].astype(np.int64)
